fixings for public list and ens adresses
